Just seperated our main aap i.e MyNeighbourhood app into their respective components.
